+++ lib/graphql/language/lexer.rb
@@ -3,6 +3,10 @@
 module GraphQL
 module Language
 module Lexer
+if !String.method_defined?(:match?)
+	using GraphQL::StringMatchBackport
+end
+
 def self.tokenize(query_string)
 	run_lexer(query_string)
 end
@@ -163,7 +167,7 @@ end
 self.graphql_lexer_en_main  = 23;
 
 def self.run_lexer(query_string)
-	data = query_string.unpack("c*")
+	data = query_string.unpack(PACK_DIRECTIVE)
 	eof = data.length
 	
 	# Since `Lexer` is a module, store all lexer state
@@ -1421,13 +1425,13 @@ def self.emit_string(ts, te, meta, block:)
 quotes_length = block ? 3 : 1
 value = meta[:data][ts + quotes_length, te - ts - 2 * quotes_length].pack(PACK_DIRECTIVE).force_encoding(UTF_8_ENCODING) || ''
 line_incr = 0
-if block && !value.length.zero?
+if block && !value.empty?
 line_incr = value.count("\n")
 value = GraphQL::Language::BlockString.trim_whitespace(value)
 end
 # TODO: replace with `String#match?` when we support only Ruby 2.4+
 # (It's faster: https://bugs.ruby-lang.org/issues/8110)
-if !value.valid_encoding? || value !~ VALID_STRING
+if !value.valid_encoding? || !value.match?(VALID_STRING)
 meta[:tokens] << token = GraphQL::Language::Token.new(
 :BAD_UNICODE_ESCAPE,
 value,