+++ spec/graphql/schema/member/scoped_spec.rb
@@ -12,6 +12,13 @@ describe GraphQL::Schema::Member::Scoped do
           items.select { |i| i.name == "Trombone" }
         elsif context[:english]
           items.select { |i| i.name == "Paperclip" }
+        elsif context[:lazy]
+          # return everything, but make the runtime wait for it,
+          # and add a flag for confirming it was called
+          ->() {
+            context[:proc_called] = true
+            items
+          }
         else
           # boot everything
           items.reject { true }
@@ -105,6 +112,7 @@ describe GraphQL::Schema::Member::Scoped do
       use GraphQL::Execution::Interpreter
       use GraphQL::Analysis::AST
     end
+    lazy_resolve(Proc, :call)
   end
 
   describe ".scope_items(items, ctx)" do
@@ -177,6 +185,25 @@ describe GraphQL::Schema::Member::Scoped do
       assert_equal ["Paperclip"], names
     end
 
+    it "works for lazy connection values" do
+      ctx = { lazy: true }
+      query_str = "
+      {
+        itemsConnection {
+          edges {
+            node {
+              name
+            }
+          }
+        }
+      }
+      "
+      res = ScopeSchema.execute(query_str, context: ctx)
+      names = res["data"]["itemsConnection"]["edges"].map { |e| e["node"]["name"] }
+      assert_equal ["Trombone", "Paperclip"], names
+      assert_equal true, ctx[:proc_called]
+    end
+
     it "is called for abstract types" do
       query_str = "
       {
@@ -193,6 +220,12 @@ describe GraphQL::Schema::Member::Scoped do
       res = ScopeSchema.execute(query_str, context: {first_letter: "T"})
       things = res["data"]["things"]
       assert_equal [{ "name" => "Trombone" }, {"designation" => "Turbine"}], things
+    end
+
+    it "works with lazy values" do
+      ctx = {lazy: true}
+      assert_equal ["Trombone", "Paperclip"], get_item_names_with_context(ctx)
+      assert_equal true, ctx[:proc_called]
     end
   end